Task8

+++ 8 - Number Divider.cpp
@@ -25,8 +25,8 @@ unsigned int BestDivider(unsigned int input) {
 }
 
 int main() {
-  constexpr unsigned int MIN = std::numeric_limits<unsigned int>::min();
-  constexpr unsigned int MAX = std::numeric_limits<unsigned int>::max();
+  constexpr unsigned long MIN = std::numeric_limits<unsigned long>::min();
+  constexpr unsigned long MAX = std::numeric_limits<unsigned long>::max();
   printf("Please enter a number to calculate all divisors and its best divisor "
          "within a range of [%u;%u]",
          MIN, MAX);
